+++ src/components/Overlay/Overlay.module.scss
@@ -1,11 +1,37 @@
 .overlay {
-  width: 100%;
-  height: 100%;
-  position: absolute;
+  margin: auto;
+  align-items: center;
+
+  display: flex;
+  width: 500px;
+  height: 500px;
+  position: fixed;
   top: 0;
   left: 0;
+  border: 1px solid red;
 
   box-sizing: border-box;
 
   background-color: var(--shadow);
+  visibility: visible;
+
+  z-index: 10;
+
+  &Open {
+    visibility: visible;
+  }
+
+  &Container {
+    position: relative;
+    z-index: 1;
+
+    display: flex;
+    flex-direction: column;
+
+    width: 500px;
+    min-height: 220px;
+    padding: 64px;
+
+    background-color: var(--white);
+  }
 }
