--- src/components/Popup/Popup.module.scss
@@ -1,62 +1,0 @@
-.modal {
-  position: fixed;
-  top: 0;
-  left: 0;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  width: 100vw;
-  height: 100vh;
-
-  z-index: 10;
-
-  &Container {
-    position: relative;
-    z-index: 1;
-
-    display: flex;
-    flex-direction: column;
-
-    width: 500px;
-    min-height: 220px;
-    padding: 64px;
-
-    background-color: var(--accent-30);
-    border: 1px solid rgb(244 244 247 / 10%);
-    border-radius: 4px;
-  }
-
-  &Close {
-    cursor: pointer;
-
-    position: absolute;
-    top: 12px;
-    right: 12px;
-
-    width: 40px;
-    height: 40px;
-    padding: 8px;
-
-    background: none;
-    border: none;
-
-    div svg {
-      width: 24px;
-      height: 24px;
-    }
-  }
-
-  &Header {
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 32px;
-
-    button {
-      cursor: pointer;
-      background-color: transparent;
-      border: none;
-    }
-  }
-}
