+++ src/components/PopupContact/PopupContact.module.scss
@@ -1,0 +1,3 @@
+.popupTitle {
+  font-family: --title-h-3-font-family;
+}
